LayoutAnimation: Enabled
