+++ Libraries/LayoutAnimation/LayoutAnimation.web.js
@@ -1,0 +1,17 @@
+/**
+ * Copyright (c) 2015-present, Alibaba Group Holding Limited.
+ * All rights reserved.
+ *
+ * Copyright (c) 2015, Facebook, Inc.  All rights reserved.
+ *
+ * @providesModule ReactLayoutAnimation
+ */
+"use strict";
+
+var LayoutAnimation = {
+  Types: {},
+  Properties: {},
+  configureNext: function configureNext() {}
+};
+
+module.exports = LayoutAnimation;
