--- Libraries/LayoutAnimation/Layoutanimation.h5.js
@@ -1,9 +1,0 @@
-"use strict";
-
-var LayoutAnimation = {
-  Types: {},
-  Properties: {},
-  configureNext: function configureNext() {}
-};
-
-module.exports = LayoutAnimation;
